Moved ViewResolver article related code from spring-mvc-java to spring-mvc-basics
