+++ spring-mvc-basics/pom.xml
@@ -36,12 +36,12 @@
             <artifactId>jstl</artifactId>
             <version>${jstl.version}</version>
         </dependency>
-        
+
         <!-- Jackson -->
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-             <version>${jackson.version}</version>
+            <version>${jackson.version}</version>
         </dependency>
 
         <dependency>
@@ -55,6 +55,18 @@
             <artifactId>json-path</artifactId>
             <scope>test</scope>
             <version>${jayway.json-path.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.rest-assured</groupId>
+            <artifactId>rest-assured</artifactId>
+            <version>${rest-assured.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.hamcrest</groupId>
+            <artifactId>hamcrest-all</artifactId>
+            <version>${hamcrest.version}</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 
@@ -72,6 +84,8 @@
         <javax.servlet-api.version>4.0.1</javax.servlet-api.version>
         <javax.jsp-api.version>2.3.3</javax.jsp-api.version>
         <jayway.json-path.version>2.4.0</jayway.json-path.version>
+        <rest-assured.version>4.0.0</rest-assured.version>
+        <hamcrest.version>1.3</hamcrest.version>
     </properties>
 
 </project>
