+++ spring-mvc-basics/src/main/java/com/baeldung/spring/web/config/WebConfig.java
@@ -3,12 +3,15 @@ package com.baeldung.spring.web.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.ClassPathResource;
 import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.view.InternalResourceViewResolver;
 import org.springframework.web.servlet.view.JstlView;
+import org.springframework.web.servlet.view.ResourceBundleViewResolver;
+import org.springframework.web.servlet.view.XmlViewResolver;
 
 @EnableWebMvc
 @Configuration
@@ -27,7 +30,23 @@ public class WebConfig implements WebMvcConfigurer {
         bean.setViewClass(JstlView.class);
         bean.setPrefix("/WEB-INF/view/");
         bean.setSuffix(".jsp");
+        bean.setOrder(2);
+        return bean;
+    }
+    
+    @Bean
+    public ViewResolver resourceBundleViewResolver() {
+        final ResourceBundleViewResolver bean = new ResourceBundleViewResolver();
+        bean.setBasename("views");
         bean.setOrder(0);
+        return bean;
+    }
+    
+    @Bean
+    public ViewResolver xmlViewResolver() {
+        final XmlViewResolver bean = new XmlViewResolver();
+        bean.setLocation(new ClassPathResource("views.xml"));
+        bean.setOrder(1);
         return bean;
     }
 }
