+++ spring-mvc-basics/src/main/java/com/baeldung/web/controller/SampleController.java
@@ -5,9 +5,20 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class SampleController {
-	@GetMapping("/sample")
+
+    @GetMapping("/sample")
     public String showForm() {
         return "sample";
+    }
+    
+    @GetMapping("/sample2")
+    public String showForm2() {
+        return "sample2";
+    }
+    
+    @GetMapping("/sample3")
+    public String showForm3() {
+        return "sample3";
     }
 
 }
