+++ spring-mvc-basics/src/main/resources/views.properties
@@ -1,0 +1,3 @@
+sample2.(class)=org.springframework.web.servlet.view.JstlView
+sample2.url=/WEB-INF/view2/sample2.jsp
+
