+++ spring-mvc-basics/src/main/webapp/WEB-INF/view2/sample2.jsp
@@ -1,0 +1,7 @@
+<html>
+<head></head>
+
+<body>
+	<h1>This is the body of the sample2 view</h1>
+</body>
+</html>
