+++ spring-mvc-basics/src/main/webapp/WEB-INF/view3/sample3.jsp
@@ -1,0 +1,7 @@
+<html>
+<head></head>
+
+<body>
+	<h1>This is the body of the sample3 view</h1>
+</body>
+</html>
