--- spring-mvc-java/src/main/resources/views.properties
@@ -1,3 +1,0 @@
-sample.(class)=org.springframework.web.servlet.view.JstlView
-sample.url=/WEB-INF/view/sample.jsp
-
