--- spring-mvc-java/src/main/resources/views.xml
@@ -1,10 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans
-        http://www.springframework.org/schema/beans/spring-beans-4.2.xsd">
- 
-    <bean id="sample" class="org.springframework.web.servlet.view.JstlView">
-        <property name="url" value="/WEB-INF/view/sample.jsp" />
-    </bean>
-    
-</beans>
